+++ botfront/imports/ui/components/synonyms/Gazette.jsx
@@ -57,7 +57,6 @@ class GazetteEditor extends React.Component {
                 ),
                 width: 100,
                 filterable: false,
-                className: 'right',
             },
         ];
     }